0.1.3 Beta
-Перенесены почти все интерфейсы в отдельные классы
- У всех интерфейсов есть айди
- У кнопок тоже есть айди

+++ core/src/com/kgc/sauw/graphic/Graphic.java
@@ -12,15 +12,15 @@ public final class Graphic {
     public static final Camera2D GAME_CAMERA;
     public static final Camera2D INTERFACE_CAMERA;
     public static final Camera2D MENU_CAMERA;
-    public static final int SCREEN_WIDTH;
-    public static final int SCREEN_HEIGHT;
+    public static final float SCREEN_WIDTH;
+    public static final float SCREEN_HEIGHT;
     static {
         SCREEN_WIDTH = Gdx.graphics.getWidth();
         SCREEN_HEIGHT = Gdx.graphics.getHeight();
         BATCH = new SpriteBatch();
         TEXTURES = new Textures();
         TEXTURES.load();
-        BLOCK_SIZE = SCREEN_WIDTH / 16;
+        BLOCK_SIZE = (int) (SCREEN_WIDTH / 16);
         GAME_CAMERA = new Camera2D();
         GAME_CAMERA.setCurrentCameraZoom(1.25f);
         INTERFACE_CAMERA = new Camera2D();
